+++ web/web/urls.py
@@ -1,0 +1,26 @@
+"""web URL Configuration
+
+The `urlpatterns` list routes URLs to views. For more information please see:
+    https://docs.djangoproject.com/en/1.11/topics/http/urls/
+Examples:
+Function views
+    1. Add an import:  from my_app import views
+    2. Add a URL to urlpatterns:  url(r'^$', views.home, name='home')
+Class-based views
+    1. Add an import:  from other_app.views import Home
+    2. Add a URL to urlpatterns:  url(r'^$', Home.as_view(), name='home')
+Including another URLconf
+    1. Import the include() function: from django.conf.urls import url, include
+    2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
+"""
+from django.conf.urls import url
+from django.contrib import admin
+
+from views import index_page, view_projects, about_page
+
+urlpatterns = [
+    url(r'^admin/', admin.site.urls),
+    url(r'^$', index_page),
+    url(r'^projects/(?P<id>\d+)/view/$', view_projects),
+    url(r'^about', about_page),
+]